Feature/calendar delete api

+++ src/main/java/com/junsu/cyr/repository/CalendarRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
+public interface CalendarRepository extends JpaRepository<Calendar, Long> {
     List<Calendar> findByDateBetween(LocalDate start, LocalDate end);
 
     @Query("select c from Calendar as c where c.date between :start and :end order by c.date desc")
